dinereserve-backend/requirements.txt
fastapi
uvicorn
motor==3.3.2
pydantic[email]
python-dotenv==1.0.0
passlib[bcrypt]
python-jose[cryptography]
email-validator
pymongo==4.6.1
scikit-learn==1.5.2
pandas==2.2.0
numpy==1.26.4
joblib==1.3.2
bcrypt==4.0.1
vaderSentiment
httpx
python-multipart
uuid